+++ src/jsx/aeft/aeft.ts
@@ -8,6 +8,8 @@ import {
 } from "../utils/samples";
 export { helloError, helloStr, helloNum, helloArrayStr, helloObj, helloVoid };
 import { dispatchTS } from "../utils/utils";
+import { applyEasing } from "./easing";
+export { applyEasing };
 
 export const helloWorld = () => {
   alert("Hello from After Effects!");